Drop the publish cron; continue in-process and adopt stale jobs
The cron ran every minute — ~1,440 invocations a day, almost all finding nothing,
for an internal tool that publishes rarely and where 96% of works finish in the
first tick. Removed, along with vercel.json, which existed only to declare it.

Continuation now uses next/server's after(), which keeps working once the response
is sent, so a large work drives itself without holding the client open and without
a scheduler. No new dependency: after() ships with Next 16.

Recovery no longer needs a sweep either. The reason a stalled job mattered was the
one-live-job-per-work index blocking republishing; start_publish_job now adopts a
job whose tick lease expired beyond a grace period, resuming from its checkpoint.
Retrying a publish is the recovery path, which is the action a person already takes
when something looks hung. Verified end to end: a job abandoned mid-artifact was
adopted on retry, kept its version label and checkpoint rather than starting over,
and completed.

Also splits the resolver into publish.query.ts / publish.mutation.ts with shared
mapping helpers, per the convention documented in apps/api-graphql/CLAUDE.md, and
corrects comments and README text that still described a cron sweep, a --dry-run
flag that no longer exists, and rows being materialized from the artifact (true of
rebuild, not of publish).
