+++ apps/api-graphql/src/graphql/resolvers.ts
@@ -18,10 +18,8 @@ import {
 import { healthQueries } from './schema/health/health.query';
 import { lookupQueries } from './schema/lookup/lookup.query';
 import { passageQueries } from './schema/passage/passage.query';
-import {
-  publishMutations,
-  publishQueries,
-} from './schema/publish/publish.resolver';
+import { publishMutations } from './schema/publish/publish.mutation';
+import { publishQueries } from './schema/publish/publish.query';
 import {
   replaceMutation,
   savePassagesMutation,
