--- apps/api-graphql/vercel.json
@@ -1,9 +1,0 @@
-{
-  "$schema": "https://openapi.vercel.sh/vercel.json",
-  "crons": [
-    {
-      "path": "/api/cron/publish-tick",
-      "schedule": "* * * * *"
-    }
-  ]
-}
